+++ lib/arachni/http.rb
@@ -561,6 +561,7 @@ class HTTP
                         else
                             @_404[path][i]['body'] ||= c_res.body
                             @_404[path][i]['rdiff'] = @_404[path][i]['body'].rdiff( c_res.body )
+                            @_404[path][i]['rdiff_words'] = @_404[path][i]['rdiff'].words.map( &:hash )
                         end
                     end
                 }
@@ -672,9 +673,22 @@ class HTTP
     end
 
     def is_404?( path, body )
+        # give the rDiff algo a shot first hoping that a comparison of
+        # refined responses will be enough to give us a clear-cut positive
         @_404[path].each do |_404|
             return true if _404['body'].rdiff( body ).hash == _404['rdiff']
         end
+
+        # if the comparison of the refinements fails, compare them based on how
+        # many words are different between them
+        @_404[path].each do |_404|
+            rdiff_body_words = _404['body'].rdiff( body ).words.map( &:hash )
+            return true if (
+                (_404['rdiff_words'] - rdiff_body_words) -
+                (rdiff_body_words - _404['rdiff_words'])
+            ).size < 25
+        end
+
         false
     end